Feature/qwen image edit (#152)
* support qwen image edit

* fix Qwen Image Edit resolution

* rope max size 4096 -> 10000

* fix

---------

+++ diffsynth_engine/pipelines/base.py
@@ -164,7 +164,7 @@ class BasePipeline:
     @staticmethod
     def generate_noise(shape, seed=None, device="cpu", dtype=torch.float16):
         generator = None if seed is None else torch.Generator(device).manual_seed(seed)
-        noise = torch.randn(shape, generator=generator, device=device).to(dtype)
+        noise = torch.randn(shape, generator=generator, device=device, dtype=dtype)
         return noise
 
     def encode_image(
